move data conversion into its own function and run in calc() phase
